+++ src/Shoelace.Style/Components/Drawer/DrawerParameters.cs
@@ -1,0 +1,13 @@
+﻿namespace Shoelace.Style.Components;
+
+/// <summary>
+/// The parameters passed into a <see cref="ShoelaceDrawer"/> instance.
+/// </summary>
+public class DrawerParameters : BaseHelperParameters
+{
+    /// <summary>
+    /// The default Drawer parameters.
+    /// This field is only intended for parameters that do not differ from their default values.
+    /// </summary>
+    internal static readonly DrawerParameters Default = [];
+}